e2e: fix split-editor reload and terminal listing flakes
Two flakes seen in the pr-18508 all-platforms run; neither is caused by
that PR.

The split-editor reload test raced the ~5s client state push timer: a
reload before the activeTabDocId push restores a stale active tab, so
the document comes back as a background tab that is never laid out, and
neither its text nor the split ever renders (flaked on all four
platforms). Reuse the #17944 fix: wait for the set_client_state push
carrying the active doc id before reloading. The helper moves from
active_tab_restore.test.ts into utils/client-state.ts so both tests
share it.

The terminal listing test required the created filename to appear twice
in terminalBuffer(), once from the echoed touch command and once from
the ls output. terminalBuffer() returns physical lines, and on macOS CI
the runner's GUID hostname makes the prompt long enough that the echoed
command wraps mid-filename, dropping one match. Type the touch argument
with a quote split so the echo can never contain the contiguous
filename, and require a single intact match, which can then only come
from the ls output (which starts at column 0 and cannot wrap).
